:operator bool): Put the original test for gobject_.stamp != 0 back in
* gtk/src/treeiter.ccg (TreeIter::operator bool): Put the original
test for gobject_.stamp != 0 back in place.  The iter_is_valid()
method provided by model implementations is intended for debugging
purposes only, and thus should not be used here (bug #142599).

* gtk/src/treemodel.hg (TreeModel::iter_is_valid): Document that
the usage of this method as abstract interface is deprecated, and
that the implementation's iter_is_valid() should be used directly.