Check for gthread-2.0 and if it exists, substitute the required compiler
* configure.in: Check for gthread-2.0 and if it exists, substitute
the required compiler flags into GTHREAD_CFLAGS.  This is necessary
to make the library code work correctly in a threaded application.
For instance, on glibc systems errno is replaced by a macro which
expands into a function call.  Actually this stuff was already in
place, but apparently got lost when glibmm was split off.

* build_shared/Makefile_build.am_fragment (all_includes): Append
$(GTHREAD_CFLAGS).