Correct misuse of container handle (bgo #586626)
* gtk/src/clipboard.{ccg,hg} (Clipboard::wait_for_targets): Create
an intermediate array of allocated C strings and transfer ownership
on return, instead of trying to returned a handle to a container
with local scope.  (Based on patch by Peter Clifton.)
(SignalProxy_TargetsReceived_gtk_callback): Correct another misuse
of the container handles similar to the above, except for the crash.
(SignalProxy_RichTextReceived_gtk_callback): Simplify the code and
plug a memory leak in the event of an exception.