Window: Added get_device_cursor().
* gdk/src/device.[hg|ccg]: Move the EventMask and GrabStatus enums here
because they are used by Device and Device is the more common header to
include. Do not include Window from here.
* gdk/src/window.hg: Added get_device_cursor(). Include device.h to use the
InputSource enum.
* tools/m4/convert_gdk.m4: Added necessary conversion.
* gdk/src/display.[hg|ccg]: Do not include device.h or devicemanager.h,
to avoid a circular include.
* gdk/src/screen.hg: Remove mention of removed gdk_screen_get_rgb_visual()
function.
* gdk/src/event.hg: Include types.h instead of window.h, because that is all
that is needed, and it avoids a circular include.
* gdk/src/pixbuf.hg: Include cairomm/surface.h instead of window.h, because
that is all that is needed, and it avoids a circular include.
* gdk/src/gdk_extra_objects.defs: Mention more objects, to improve the
documentation.
* gdk/gdmm/general.h: Include rgba.h and rectangle.h now that they are not
included indirectly.