Gtk::Application: Try to make this a replacement for Gtk::Main.
* gtk/src/application.[hg|ccg]: Add a copy of init_gtkmm_internals() from
Gtk::Main's main.cc and call it from the constructor, so that
Gtk::Application can be a replacement for Gtk::Main.
Add a constructor that takes argc and argv, so we can completely
initialize GTK+ and gtkmm, because doing it in run() will be too late for
some code: See bug #639925.
Add a run() that does not take argc, argv, for when you have supplied them
to the constructor instead.
Add a run() that takes a window, like Gtk::Main::run(), so it can return
when the window closes. However, see bug #639931 about it really responding to
destruction, not closing.
We show the window in an activate signal handler, because GTK+ complains if
we don't have such a handler. See bug #640042.