Regenerated enum .defs files.
* gdk/src/gdk_enums.defs:
* gtk/src/gtk_enums.defs:
Regenerated with enums.pl
Apparently we did not regenerate these recently before 3.0.0.
There is a (then late) change in GdkDragProtocol that actually changes
the enum values. This is technically an ABI break, but we would be passing
the wrong values to the C functions, leading to already incorrect behaviour,
so this can only be an improvement.
This also removes other enums from the .defs that we do not use any more in
the actual API.