Application: Avoid the need to use Gtk::Main.
* gtk/src/application.[hg|cc]: Add an init_gtkmm() method, for use in
the create() methods of derived Application classes. This is necessary
to allow the Application constructor to succeed, because it uses
g_object_new().
create(): Use it here too, though I wonder if anyone would actually use
Gtk::Application without deriving.
This lets developers no use Gtk::Main at all, as long as they call their
Application's create() method at the same early point in their code.

This does not allow the developer to defer GTK+ initialization to when they
call g_application_run(), and even to avoid GTK+ initialization compeletely
in the local instance, as is possible with the C API, but we did not allow
that before either.