Update to be more compatible with glibmm 2.47.6 and later
* gdk/src/types.hg: Glib::ScopedPtr has been deprecated. To be able to build
gtkmm with GLIBMM_DISABLE_DEPRECATED, copy Glib::ScopedPtr to Gdk::ScopedPtr.
* gdk/src/types.ccg:
* gtk/src/clipboard.ccg:
* gtk/src/colorselection.ccg:
* gtk/src/rc.ccg:
* gtk/src/widget.ccg:
* demos/gtk-demo/example_dialog.cc: Use Gdk::ScopedPtr instead of
Glib::ScopedPtr.
* gtk/src/menushell.ccg: Glib::Refptr::operator bool() has been made explicit.
It's no longer used in item != 0, where item is a RefPtr. Change the
expression to the equivalent !!item.

Even with these changes it's not unproblematic to build gtkmm2 with a new
version of gmmproc, because _CLASS_BOXEDTYPE and _CLASS_GOBJECT generate
move operators but gtkmm2 does not require a C++11 compiler.