Refactor Visual Studio 2010 projects
This consolidates common items into the property sheets, and makes the
header/lib search more like the project files that is provided by the GTK+
stack, so that we can simplify the project files, to ease future
maintenance  This also generates the .pdb files for all builds, to ease
debugging.

This also simplifies the build with later Visual Studio versions so that
the .dll/.lib/.pdb  filename can be updated by just updating the
gtkmm-version-paths.props property sheet.