Visual Studio builds: Require Visual Studio 2017
The latest gtkmm-4.x will require C++-14 features that are supported
only by Visual Studio 2017 and later, so update the projects
accordingly.  Also update mentions of earlier Visual Studio versions to
refer to this Visual Studio version.

For the C++ dependencies of gtkmm, use the property sheets to record the
version names so that we can update them more easily in the future if
the need should arise.