FileChooserDialog: Allow constructing with use-header-bar
This property is handy e.g. for ensuring consistent UI across platforms.
This is a construct-only property, so we must specifically implement it.
Doing so is free, by adding it as a defaulted argument to our remaining
ctors, with the default value of false equivalent to what we got before.

https://bugzilla.gnome.org/show_bug.cgi?id=780004