Demo: Don't pass padding to Box::pack_start().
The parameter was removed from the gtkmm API in this commit:
https://git.gnome.org/browse/gtkmm/commit/gtk/src/box.hg?id=a53eea323ab9f9b3d228d72c6c13998b73f58ab3
in response to this GTK+ API change:
https://git.gnome.org/browse/gtk+/commit/?id=f45dadadd00d892ba16dc081240987425bc32346
but the compiler was implicitly converting the enum and the int to bool,
bool, to call pack_start(widget, bool, bool) instead, which is not what
was wanted.
I have made changes to the demo code to match the C commit above, more or less, though
they don't make much sense.