demos: Adapt to changed Box::pack_start() behaviour.
pack_start(EXPAND_WIDGET) or pack_start(EXPAND_PADDING), or the same
with pack_end(), sets the child's hexpand and vexpand to true.
But that also affects the parent widget (the old fill/expand GTK+ API
didn't do that), so sometimes we need to unset that.

Well, that explains most of these changes. I found the need for these
changes by comparing the appearance (and behaviour when resizing) of
gtkmm-demo for both gtkmm 3 and gtkmm 4.