Add a specialization for Glib::Value<::Cairo::RefPtr<T>>
* gdk/gdkmm/value_cairo.[h|cc]: New files. Add a specialization for
Glib::Value<::Cairo::RefPtr<T>>. This specialization can't be put in glibmm,
because glibmm does not depend on cairomm. It can't be put in cairomm,
because cairomm does not depend on glibmm. So it's added here. It's needed
for wrapping properties with a Cairo::RefPtr<T> value.
* gdk/gdkmm/filelist.am: Add value_cairo.[h|cc].
* gtk/src/cellrendererpixbuf.hg: Include value_cairo.h. It's needed for
property_surface(). Bug 788513