Gtk::Object::_release_c_instance(): Don't release if in a container
Don't release the C instance, if it's a widget in a container widget.
Just steal the GQuark that points from the C instance to the C++ wrapper,
and accept that a new wrapper is later created, if necessary.
Nowadays it's difficult to know which container the widget has been added to.
It's not necessarily the one returned by gtk_widget_get_parent().
Bug 773642 comments 20-22. See also gtk+ bug 786048 comment 5.