gendef: Skip CRT symbols that are implemented inline
As a result to the move of UCRT for Visual Studio 2015 and later, some
of the CRT functions are no longer done in the CRT DLL, but are instead
implemented directly in the CRT headers.  We need to filter out these
symbols so that we can link the gtkmm DLL properly.

Unfortunately we cannot just link to legacy_stdio_definitions.lib to
resolve this, so we just have to strip these symbols out from the
generated gtkmm.def.