build: Add NMake Makefiles for Visual Studio builds
This adds NMake Makefiles for Visual Studio 2013 and later, which will
replace the Visual Studio projects.  The main motivation for this is to
ease maintenace on the Visual Studio build files, especially as sources
are often added or removed, making the project files out-of-date as they
are not generated during 'dist make'.  This will share the various
filelist.am's so that any additions or removal of sources can be
reflected upon the Visual Studio builds immediately.

Since this is the C++-11 version of gtkmm which supports both Visual
Studio 2015 and 2017, we want to name the DLL and .lib to be like
<library>-vc140-3_0.[dll.lib] or <library>-vc140-d-3_0.[dll|lib] since
they both link to the v140 Windows C/C++ runtime DLLs, and thus these
DLLs and .lib's are advertised to be interoperable, as supported by
Microsoft.

Note that the master branch, which requires C++-17, is not changed
as Visual Studio 2017 is really required to build things there.

Note that as some code in the demos require compiler features that is
only available in Visual Studio 2015 or later, we only build the demos
if Visual Studio 2015 or later is used.