build: Add NMake Makefiles for Visual Studio builds
This adds NMake Makefiles for Visual Studio 2017 and later, which will
replace the Visual Studio projects.  The main motivation for this is to
ease maintenace on the Visual Studio build files, especially as sources
are often added or removed, making the project files out-of-date as they
are not generated during 'dist make'.  This will share the various
filelist.am's so that any additions or removal of sources can be
reflected upon the Visual Studio builds immediately.