MSVC_NMake/config-msvc.mak: Use $(PDBVER) in DLL names
Since this is the C++-11 version of gtkmm which supports both Visual
Studio 2015 and 2017, we want to name the DLL and .lib to be like
<library>-vc140-3_0.[dll.lib] or <library>-vc140-d-3_0.[dll|lib] since
they both link to the v140 Windows C/C++ runtime DLLs, and thus these
DLLs and .lib's are advertised to be interoperable, as supported by
Microsoft.

Note that the master branch, which requires C++-17, is not changed
as Visual Studio 2017 is really required to build things there.