Builder: Add get_object<Derived>(name) overloads
Allow getting and casting to a subclass of Glib::Object in one line,
thus making it nicer to deal with non-widget objects in Builder files.

The non-template and template member functions act as distinct
overloads, meaning the new ones can be named get_object() also.

Add a Gtk::Adjustment to tests/builder, and check its refcounts are OK.

https://gitlab.gnome.org/GNOME/gtkmm/issues/43