Builder: Add get_object<Derived>(name) overloads
Allow getting and casting to a subclass of Glib::Object in one line,
thus making it nicer to deal with non-widget objects in Builder files.

The non-template and template member functions act as distinct
overloads, meaning the new ones can be named get_object() also.

Add a Gtk::Adjustment to tests/builder, and check its refcounts are OK.
Also add a g_assert that (A) ensures that the test compiles, i.e. we got
the right type in the RefPtr and (B) checks we get the right :value out.
In case (B) seems frivolous, I did see it not working due to prop order!
Adjustment:value needs set after :lower|:upper as it is clamped by them.

While here, g_assert() the mm pointers/RefPtrs just to be fully paranoid

https://gitlab.gnome.org/GNOME/gtkmm/issues/43