Builder: Add returning vers of get_widget|_derived
To avoid the cumbersome two-step process of declaring a pointer and then
passing it to get_widget[_derived]() for use as an output argument, add
overloads that return the pointer directly. Of course, now we have no
arg from which to deduce the type, so pass that as a template argument.

https://gitlab.gnome.org/GNOME/gtkmm/issues/43