tools: _WRAP_CHILD_PROPERTY: Check the data type
tools/m4/child_property.m4: _WRAP_CHILD_PROPERTY generates a static_assert()
that checks if the data type is acceptible. The generation of the
static_assert() can be suppressed with the new parameter no_type_check.

For this to work, the code must be generated with a glibmm/gmmproc version
from 2019-01-31 or newer.