Build: Build gdkmm into gtkmm
In the autotools builds, GDKMM is no longer built as a standalone DLL
but is instead part of the GTKMM DLL.  Make it the case in the Visual
Studio builds as well.  This makes things in line with what happens in
GTK master.

Also remove all traces of gdkmm.rc, since that is no longer used.