NMake Makefiles: Add rules to generate sources
This way, we are now able to build directly from a GIT checkout by first
running the 'prep-git-build' target before running the NMake build
itself.

This will require PERL (with XML::Parser) and some common UNIX utilities
such as 'cp' and 'rm' in order to run gmmproc from glibmm.  Note that
the m4 files for atkmm and pangomm used for their code generation are
required as well.