NMake Makefiles: Silence some warnings
Enable /EHsc so that we do not receive warning C4577 for constexpr
usage.

Also disable warnings C4251 and C4275 as they relate to whether we are
using __declspec(dllimport) for the public symbols in the glibmm
headers, when building against glibmm 2.64.0 or later.  Since we know
what we are indeed going to use __declspec(dllimport) as we build against
glibmm 2.64.0 or later, we can just ignore those warnings.