NMake Makefiles: Support linking to Meson-built -mm deps
This makes it easier for people that want to build dependencies with
Meson as far as possible, by adding an USE_MESON_LIBS NMake option.

Note that by using USE_MESON_LIBS at this time of writing, libsigc++,
glibmm and pangomm should be built by Meson, and atkmm and cairomm
should also link to the Meson-built variants of glibmm and libsigc++ as
well (as atkmm and cairomm do not yet have Meson build files).