NMake Makefiles: Support generating sources with gmmproc
This  way, we can support building directly from GIT checkouts, after
running the 'prep-git-build' target.  This will require the latest .m4
files and gmmproc from glibmm, atkmm and pangomm master, along with PERL
with XML::Parser and some common *nix utilities.

Update the build info so that people know what options they may need to
make use of this new build feature.

At the same time, improve the 'clean' target so that we can reduce the
spew when the test programs are not built.

Also add MSVC_NMake/pkg-ver.mak into .gitignore, as it is generated.