gdk/gdkmm/*.h: Mark functions with GDKMM_API
If gmmproc is new enough to generate the full sources from the .hg files
with GDKMM_API applied to all applicable places, we can use this to
export symbols without the need to use gendef.exe.

The build-time check to check if gmmproc is new enough will be in an
upcoming commit.