docs/reference/images/stock: Remove .svg files
There are about 200 unused .svg files. Remove them. Stock items are
deprecated. They are removed in gtk4 and gtkmm4.

When Meson support is added, and tarballs are created with 'ninja dist',
all files stored in git will be included in the tarballs unless
removed by a custom command. It's easier to remove unused files
from the git repo.