NMake Makefiles: Use /utf-8 if available
This flag is provided on Visual Studio 2015 and later, so that it will help to
work around unicode handling issues that occur when building on East Asian
locales, that is demonstrated by error/warning C4819.

Disabling warning C4828 is meant for gendef.exe, which comes as a consequence
of using /utf-8.