Meson: Clean-up and fix dependency search
Since we have Meson build files for the C++ dependencies, just use the normal
dependency search via pkg-config for all the -mm dependencies.  This also
includes looking for atkmm, unless we request not to.

We do, however, want to look for gtk+-unix-print-3.0 only if we are not
building on Windows (at least), becuase there is no way that it will be there
on Windows.

Also clean up the parts where we decide whether to build gdkmm/gtkmm as a
shared library directly without using gendef.exe.