Meson: Ignore more warnings on Visual Studio
We can safely ignore warnings C4251, C4273 and C4275 as we are sure that we
want to use __declspec(dllimport) on the libraries that we depend upon, and
even with the C++11 versions, we are basically tied to the same CRT as Visual
Studio 2015 or later is required to build the C++11 branches of the the -mm
dependencies fully.

We can also ignore warning C4805 since a 'gboolean' is defined to mimic a
'bool'.