Meson: Fix linking demos and tests on Visual Studio
GTK (and therefore gtkmm) applications require a linker flag
'-entry:mainCRTStartup' if building with 'gui_app' is set to be true (which
means, we are using /subsystem:windows in the linker flags on Visual Studio)
unless a WinMain() is defined in the individual application's sources.

This fixes the build of such programs by adding this linker flag where it is
needed.