NMake Makefiles: Stop using gendef.exe
Since the code is ready to use __declspec(dllexport) to export the symbols from
gdkmm and gtkmm, stop building and using gendef.exe.  Also make sure that we
generate and build the gdkmm sources and headers before we attempt to generate
or build the gtkmm sources and headers.

gendef.exe will be dropped in the next commit