NMake Makefiles: Fix 32-bit Visual Studio 2017 builds
We need to disable /EHsc in the compiler flags when building gtkmm-demo,
because it turns out that the linking will fail with an internal compiler error
like on Visual Studio 2015 32-bit builds.

The x64 builds on Visual Studio 2015 and later and 32-bit builds on Visual
Studio 2019 build and run as expected.