NMake Makefiles: Use toolset version in DLL/.lib naming
Instead of using the Visual Studio version ('vc150' for Visual Studio 2017),
use the toolset version ('vc141' for Visual Studio 2017 and 'vc142' for Visual
Studio 2019, as defined by Microsoft) so that we are more in-line with what
is now done in the gtkmm-3-24 branch.

If using the old naming convention is desired, an NMake command line option
'USE_COMPAT_LIBS' have been added-note that this will require glibmm, cairomm,
pangomm, atkmm and libsigc++ to be built with that option enabled.