NMake Makefiles: Fix build from release tarballs
Make sure that we do not needlessly re-generate [gdk|gtk]mm[config.h|.rc], and
ensure when they are generated, they will always contain the needed version
info in them.

Remove the gdkmm.rc target since gdkmm is now part of the main gtkmm library,
not a separate library of its own.

Also streamline the process so that it is no longer necessary to run the
'prep-git-build' target; instead, this file generating process is carried out
when necessary when running the 'all' target.