Add Gtk::ColumnView
Add bindings for ColumnView and its help class ColumnViewColumn. Also
include a binding for Gtk::Sorter, which maybe isn't complete: the
Sorter::compare method and vfunc have gpointers as arguments to raw Gtk
objects. I'm not sure a proper gtkmm wrap there is a good idea, because
of perfomance reasons.