Gdk: Add [Drag,Popup,Toplevel]SurfaceImpl
[Drag,Popup,Toplevel]SurfaceImpl don't correspond to C classes. They can
wrap any C object that derives from GdkSurface and implements
GdkDragSurface, GdkPopup and GdkToplevel, respectively. Such C classes
in GTK are not public. They can't be wrapped in the usual way.

Surface_Class::wrap_new() can wrap a C object in a Surface, DragSurfaceImpl,
PopupSurfaceImpl or ToplevelSurfaceImpl depending on which interface,
if any, the C object implements.

This is similar to DeviceWithPad.