MSVC_NMake/create-lists-msvc.mak: Work around command length limits
The list of .hg sources in gtk/src grew too long for the NMake Makefiles
to handle, so work around it by excluding the file name extensions by
omitting them when generating the list of object/.hg files dynamically,
which should keep us quite safe in that regard.

This will fix the build with the NMake Makefiles, with the fixes for
issue #85.