treeiter—Delete TreeValueProxy copy-assign/warning
The user-provided, unimplemented operator=(TreeValueProxy const&)
existed to prevent the class being copied, but since a long time C++
provides a better way to do that: declaring the operator as deleted.

Doing this also avoids warnings about the implicitly declared default
copy constructor, which is deprecated due to the previously user-given
operator=, and since the latter is now deleted the warnings are avoided.

https://gitlab.gnome.org/GNOME/gtkmm/-/issues/94