gtk/gtkmm/*.h: Export class selectively as needed
We want to avoid exporting classes that contain a std::vector<...> or
std::unique_pointer<...> member, so don't export these class as a whole but
export its methods individually.

Also, remove an extraneous GTKMM_API in listviewtext.h.

This will eliminate warnings C4251 and C4275 and avoid having the built code
depend on a specific STL and compiler version