NMake Makefiles: Some cleanups and improvements
Drop the ignores for warnings C4251, C4273 and C4275 from the warnings that we
want to ignore, as the code is now free of items that trigger those warnings.

Also, reorganize the compiler flags for gtkmm-demo as it is actually the /GL
flag that triggered the internal compiler error when it is used with /EHsc on
32-bit Visual Studio 2015 and 2017 builds.  Likewise, disable the /LTCG linker
option when building Visual Studio 2015 and 2017 32-bit builds.