meson.build: Fix builds with Vulkan-enabled GTK
Look for the Vulkan SDK so that we can find its headers, which is needed
when compiling the sources against a GTK-4.x build that has the Vulkan GSK
renderer enabled for the build.  We do not link to the Vulkan libraries to
build gtkmm, so for now we do not add a dependency in gtkmm's pkg-config files,
and Meson may have found Vulkan internally via methods other than pkg-config.