meson/MSVC: Apply /wd4828 when building gendef.exe only
This warning only applies when compiling gendef.cc, when '/utf-8' is also being
used, so only apply that compiler flag there.  Move the '/utf-8' to be with the
other warning-related compiler flags for consistency's sake.

Also, only build gendef.exe if it is really needed, i.e., glibmm's gmmproc is
not able to produce headers that can export gtkmm's symbols via compiler
directives.