SignalListItemFactory: Add signals taking GObjects
so that the factory is usable if it's producing instances of a type
other than Gtk::ListItem, e.g if it is a ListView::header-factory.

Rather than going to the lengths of defeating the default documentation,
I just prepend notes on the meaning of the two sets of signals, to the
generated documentation.